Separate cloudflare/wrangler code into its own file
